+++ src/Regressors/Adaline.php
@@ -22,12 +22,16 @@ use Rubix\ML\NeuralNet\Layers\Placeholder1D;
 use Rubix\ML\NeuralNet\FeedForward;
 use Rubix\ML\NeuralNet\Optimizers\Adam;
 use Rubix\ML\NeuralNet\Optimizers\Optimizer;
+use Rubix\ML\NeuralNet\Snapshot;
+use Rubix\ML\CrossValidation\Metrics\Metric;
+use Rubix\ML\CrossValidation\Metrics\RMSE;
 use Rubix\ML\Online;
 use Rubix\ML\Persistable;
 use Rubix\ML\RanksFeatures;
 use Rubix\ML\Specifications\DatasetHasDimensionality;
 use Rubix\ML\Specifications\DatasetIsLabeled;
 use Rubix\ML\Specifications\DatasetIsNotEmpty;
+use Rubix\ML\Specifications\EstimatorIsCompatibleWithMetric;
 use Rubix\ML\Specifications\LabelsAreCompatibleWithLearner;
 use Rubix\ML\Specifications\SamplesAreCompatibleWithEstimator;
 use Rubix\ML\Specifications\SpecificationChain;
@@ -37,8 +41,11 @@ use Rubix\ML\Verbose;
 
 use function count;
 use function get_object_vars;
+use function is_dir;
 use function is_nan;
 use function number_format;
+use function sys_get_temp_dir;
+use function uniqid;
 
 /**
  * Adaline
@@ -95,12 +102,40 @@ class Adaline implements Estimator, Learner, Online, RanksFeatures, Verbose, Per
     protected float $minChange;
 
     /**
+     * The number of epochs to train before evaluating the model with the holdout set.
+     *
+     * @var int
+     */
+    protected int $evalInterval;
+
+    /**
+     * The number of epochs without improvement in the validation score to wait before considering an early stop.
+     *
+     * @var positive-int
+     */
+    protected int $window;
+
+    /**
+     * The proportion of training samples to use for validation and progress monitoring.
+     *
+     * @var float
+     */
+    protected float $holdOut;
+
+    /**
      * The function that computes the loss associated with an erroneous
      * activation during training.
      *
      * @var RegressionLoss
      */
     protected RegressionLoss $costFn;
+
+    /**
+     * The metric used to score the generalization performance of the model during training.
+     *
+     * @var Metric
+     */
+    protected Metric $metric;
 
     /**
      * The underlying neural network instance.
@@ -117,12 +152,30 @@ class Adaline implements Estimator, Learner, Online, RanksFeatures, Verbose, Per
     protected ?array $losses = null;
 
     /**
+     * The validation scores at each epoch from the last training session.
+     *
+     * @var float[]|null
+     */
+    protected ?array $scores = null;
+
+    /**
+     * The file path to store the snapshot on disk during training.
+     *
+     * @var string|null
+     */
+    protected ?string $snapshotPath = null;
+
+    /**
      * @param int $batchSize
      * @param Optimizer|null $optimizer
      * @param float $l2Penalty
      * @param int $epochs
      * @param float $minChange
+     * @param int $evalInterval
+     * @param int $window
+     * @param float $holdOut
      * @param RegressionLoss|null $costFn
+     * @param Metric|null $metric
      * @throws InvalidArgumentException
      */
     public function __construct(
@@ -131,7 +184,11 @@ class Adaline implements Estimator, Learner, Online, RanksFeatures, Verbose, Per
         float $l2Penalty = 1e-4,
         int $epochs = 1000,
         float $minChange = 1e-4,
-        ?RegressionLoss $costFn = null
+        int $evalInterval = 3,
+        int $window = 5,
+        float $holdOut = 0.1,
+        ?RegressionLoss $costFn = null,
+        ?Metric $metric = null
     ) {
         if ($batchSize < 1) {
             throw new InvalidArgumentException('Batch size must be'
@@ -153,12 +210,35 @@ class Adaline implements Estimator, Learner, Online, RanksFeatures, Verbose, Per
                 . " greater than 0, $minChange given.");
         }
 
+        if ($evalInterval < 1) {
+            throw new InvalidArgumentException('Eval interval must be'
+                . " greater than 0, $evalInterval given.");
+        }
+
+        if ($window < 1) {
+            throw new InvalidArgumentException('Window must be'
+                . " greater than 0, $window given.");
+        }
+
+        if ($holdOut < 0.0 or $holdOut > 0.5) {
+            throw new InvalidArgumentException('Hold out ratio must be'
+                . " between 0 and 0.5, $holdOut given.");
+        }
+
+        if ($metric) {
+            EstimatorIsCompatibleWithMetric::with($this, $metric)->check();
+        }
+
         $this->batchSize = $batchSize;
         $this->optimizer = $optimizer ?? new Adam();
         $this->l2Penalty = $l2Penalty;
         $this->epochs = $epochs;
         $this->minChange = $minChange;
+        $this->evalInterval = $evalInterval;
+        $this->window = $window;
+        $this->holdOut = $holdOut;
         $this->costFn = $costFn ?? new LeastSquares();
+        $this->metric = $metric ?? new RMSE();
     }
 
     /**
@@ -202,7 +282,11 @@ class Adaline implements Estimator, Learner, Online, RanksFeatures, Verbose, Per
             'l2 penalty' => $this->l2Penalty,
             'epochs' => $this->epochs,
             'min change' => $this->minChange,
+            'eval interval' => $this->evalInterval,
+            'window' => $this->window,
+            'hold out' => $this->holdOut,
             'cost fn' => $this->costFn,
+            'metric' => $this->metric,
         ];
     }
 
@@ -230,6 +314,7 @@ class Adaline implements Estimator, Learner, Online, RanksFeatures, Verbose, Per
         foreach ($this->losses as $epoch => $loss) {
             yield [
                 'epoch' => $epoch,
+                'score' => $this->scores[$epoch] ?? null,
                 'loss' => $loss,
             ];
         }
@@ -246,6 +331,16 @@ class Adaline implements Estimator, Learner, Online, RanksFeatures, Verbose, Per
     }
 
     /**
+     * Return the validation score at each epoch from the last training session.
+     *
+     * @return float[]|null
+     */
+    public function scores() : ?array
+    {
+        return $this->scores;
+    }
+
+    /**
      * Return the underlying neural network instance or null if not trained.
      *
      * @return FeedForward|null
@@ -253,6 +348,21 @@ class Adaline implements Estimator, Learner, Online, RanksFeatures, Verbose, Per
     public function network() : ?FeedForward
     {
         return $this->network;
+    }
+
+    /**
+     * Set the file path to store the snapshot on disk during training.
+     *
+     * @param string|null $path
+     * @throws InvalidArgumentException
+     */
+    public function setSnapshotPath(?string $path) : void
+    {
+        if (isset($path) and is_dir($path)) {
+            throw new InvalidArgumentException('Snapshot path must be to a file, folder given.');
+        }
+
+        $this->snapshotPath = $path;
     }
 
     /**
@@ -305,12 +415,31 @@ class Adaline implements Estimator, Learner, Online, RanksFeatures, Verbose, Per
             $this->logger->info("{$numParams} trainable parameters");
         }
 
+        [$testing, $training] = $dataset->randomize()->split($this->holdOut);
+
+        [$minScore, $maxScore] = $this->metric->range()->list();
+
+        $bestScore = $minScore;
+        $bestEpoch = $numWorseEpochs = 0;
+        $loss = 0.0;
+        $score = $snapshot = null;
         $prevLoss = INF;
 
-        $this->losses = [];
+        $snapshotPath = $this->snapshotPath;
+
+        if (!$snapshotPath) {
+            $snapshotPath = sys_get_temp_dir() . '/rubixml-snapshot-' . uniqid() . '.dat';
+        }
+
+        if ($testing->empty() and $this->logger) {
+            $this->logger->notice('Insufficient validation data, '
+                . 'some features are disabled');
+        }
+
+        $this->scores = $this->losses = [];
 
         for ($epoch = 1; $epoch <= $this->epochs; ++$epoch) {
-            $batches = $dataset->randomize()->batch($this->batchSize);
+            $batches = $training->randomize()->batch($this->batchSize);
 
             $loss = 0.0;
 
@@ -324,16 +453,6 @@ class Adaline implements Estimator, Learner, Online, RanksFeatures, Verbose, Per
 
             $this->losses[$epoch] = $loss;
 
-            if ($this->logger) {
-                $lossDirection = $loss < $prevLoss ? '↓' : '↑';
-
-                $message = "Epoch: $epoch, "
-                    . "{$this->costFn}: $loss, "
-                    . "Loss Change: {$lossDirection}{$lossChange}";
-
-                $this->logger->info($message);
-            }
-
             if (is_nan($loss)) {
                 if ($this->logger) {
                     $this->logger->warning('Numerical under/overflow detected');
@@ -346,11 +465,68 @@ class Adaline implements Estimator, Learner, Online, RanksFeatures, Verbose, Per
                 break;
             }
 
+            $evalThisStep = $epoch % $this->evalInterval === 0 && !$testing->empty();
+
+            if ($evalThisStep) {
+                $predictions = $this->predict($testing);
+
+                $score = $this->metric->score($predictions, $testing->labels());
+
+                $this->scores[$epoch] = $score;
+            }
+
+            if ($this->logger) {
+                $message = "Epoch: $epoch, {$this->costFn}: $loss";
+
+                if ($evalThisStep) {
+                    $message .= ", {$this->metric}: $score";
+                }
+
+                $this->logger->info($message);
+            }
+
+            if ($evalThisStep) {
+                if ($score >= $maxScore) {
+                    break;
+                }
+
+                if ($score > $bestScore) {
+                    $bestScore = $score;
+                    $bestEpoch = $epoch;
+
+                    if ($snapshot) {
+                        $snapshot->destroy();
+                    }
+
+                    $snapshot = Snapshot::take($this->network, $snapshotPath);
+
+                    $numWorseEpochs = 0;
+                } else {
+                    ++$numWorseEpochs;
+                }
+
+                if ($numWorseEpochs >= $this->window) {
+                    break;
+                }
+            }
+
             if ($lossChange < $this->minChange) {
                 break;
             }
 
             $prevLoss = $loss;
+        }
+
+        if ($snapshot) {
+            if (end($this->scores) < $bestScore or is_nan($loss)) {
+                $snapshot->restore();
+
+                if ($this->logger) {
+                    $this->logger->info("Model state restored to epoch $bestEpoch");
+                }
+            }
+
+            $snapshot->destroy();
         }
 
         if ($this->logger) {
@@ -413,7 +589,12 @@ class Adaline implements Estimator, Learner, Online, RanksFeatures, Verbose, Per
     {
         $properties = get_object_vars($this);
 
-        unset($properties['losses'], $properties['logger']);
+        unset(
+            $properties['losses'],
+            $properties['scores'],
+            $properties['logger'],
+            $properties['snapshotPath']
+        );
 
         return $properties;
     }