+++ src/Regressors/MLPRegressor.php
@@ -367,7 +367,7 @@ class MLPRegressor implements Estimator, Learner, Online, Verbose, Persistable
      */
     public function setSnapshotPath(?string $path) : void
     {
-        if (isset($path) && is_dir($path)) {
+        if (isset($path) and is_dir($path)) {
             throw new InvalidArgumentException('Snapshot path must be to a file, folder given.');
         }
 
@@ -470,7 +470,9 @@ class MLPRegressor implements Estimator, Learner, Online, Verbose, Persistable
                 break;
             }
 
-            if ($epoch % $this->evalInterval === 0 && !$testing->empty()) {
+            $evalThisStep = $epoch % $this->evalInterval === 0 && !$testing->empty();
+
+            if ($evalThisStep) {
                 $predictions = $this->predict($testing);
 
                 $score = $this->metric->score($predictions, $testing->labels());
@@ -481,14 +483,14 @@ class MLPRegressor implements Estimator, Learner, Online, Verbose, Persistable
             if ($this->logger) {
                 $message = "Epoch: $epoch, {$this->costFn}: $loss";
 
-                if (isset($score)) {
+                if ($evalThisStep) {
                     $message .= ", {$this->metric}: $score";
                 }
 
                 $this->logger->info($message);
             }
 
-            if (isset($score)) {
+            if ($evalThisStep) {
                 if ($score >= $maxScore) {
                     break;
                 }
@@ -498,7 +500,7 @@ class MLPRegressor implements Estimator, Learner, Online, Verbose, Persistable
                     $bestEpoch = $epoch;
 
                     if ($snapshot) {
-                        $snapshot->clean();
+                        $snapshot->destroy();
                     }
 
                     $snapshot = Snapshot::take($this->network, $snapshotPath);
@@ -511,8 +513,6 @@ class MLPRegressor implements Estimator, Learner, Online, Verbose, Persistable
                 if ($numWorseEpochs >= $this->window) {
                     break;
                 }
-
-                unset($score);
             }
 
             if ($lossChange < $this->minChange) {
@@ -531,7 +531,7 @@ class MLPRegressor implements Estimator, Learner, Online, Verbose, Persistable
                 }
             }
 
-            $snapshot->clean();
+            $snapshot->destroy();
         }
 
         if ($this->logger) {